Create DataIngestion.py
This code defines two classes, DataIngestion and OptimizeSpeed, that can be used to extract a dataset from a table in a Microsoft SQL Server database using the pyodbc package and the tensorflow package. The DataIngestion class takes in the necessary credentials to connect to the SQL server and contains a method extract_data to extract the data from the specified table. The OptimizeSpeed class takes in an object of the DataIngestion class and contains a method optimize to optimize the speed of the data extraction process. This code is designed using object oriented design principles, and is commented and documented according to the Pep-8 style guide.